requirements.txt
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Faker==8.1.0
idna==2.10
install==1.3.4
kafka-python==2.0.2
pep8==1.7.1
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
